Add react to devDependencies
I was too enthusiastic about peerDependencies and DRYness. We need react in devDependencies in order to run tests. We specify it as a peerDependency because we mark it as an external and don't bundle it.

When using the package using yarn link for local development, the host app will complain about not being able to resolve react because it was never installed.