Change way we determine running mode
Previously, we default to standalone until the iframe resize onReady callback was called, which we will then set to embedded.

This causes a flash of standalone mode when navigating from a non-react page to the react page (going from /about to /) which to the naked eye looked like the iframe was resizing.

This change makes it explicit that you are running in embedded or standalone mode, set on the AppProvider. embedded is true by default.

Along with this change, we set the parent in a more robust way.